Extract business logic from commands into library modules
Move reusable logic out of commands/ into top-level domain packages
(deploy/, provisioning/, benchmark/, report/) so it can be imported
programmatically. Commands layer now contains only CLI handlers.
