+++ deplodock/benchmark/__init__.py
@@ -1,0 +1,36 @@
+"""Benchmark library: tracking, config, logging, workload, tasks, execution."""
+
+from deplodock.benchmark.tracking import (
+    compute_code_hash,
+    create_run_dir,
+    write_manifest,
+    read_manifest,
+)
+from deplodock.benchmark.config import load_config, validate_config, _expand_path
+from deplodock.benchmark.bench_logging import setup_logging, _get_group_logger
+from deplodock.benchmark.workload import (
+    extract_benchmark_results,
+    _parse_max_model_len,
+    run_benchmark_workload,
+)
+from deplodock.benchmark.tasks import enumerate_tasks, _task_meta
+from deplodock.benchmark.execution import run_execution_group, _run_groups
+
+__all__ = [
+    "compute_code_hash",
+    "create_run_dir",
+    "write_manifest",
+    "read_manifest",
+    "load_config",
+    "validate_config",
+    "_expand_path",
+    "setup_logging",
+    "_get_group_logger",
+    "extract_benchmark_results",
+    "_parse_max_model_len",
+    "run_benchmark_workload",
+    "enumerate_tasks",
+    "_task_meta",
+    "run_execution_group",
+    "_run_groups",
+]
