+++ deplodock/commands/vm/__init__.py
@@ -1,57 +1,7 @@
 """VM lifecycle management: create/delete cloud GPU instances."""
 
-import subprocess
-import sys
-import time
-
-
-def wait_for_ssh(host, username, ssh_port, ssh_key_path, timeout=120, interval=5):
-    """Poll SSH connectivity until success or timeout.
-
-    Uses plain ssh (not gcloud) for provider-agnostic SSH readiness check.
-
-    Returns:
-        True if SSH connected, False on timeout.
-    """
-    address = f"{username}@{host}" if username else host
-    elapsed = 0
-    while elapsed < timeout:
-        rc = subprocess.run(
-            ["ssh", "-o", "StrictHostKeyChecking=no", "-o", "UserKnownHostsFile=/dev/null",
-             "-o", "BatchMode=yes", "-o", "ConnectTimeout=5",
-             "-i", ssh_key_path, "-p", str(ssh_port),
-             address, "true"],
-            capture_output=True,
-        ).returncode
-        if rc == 0:
-            return True
-        time.sleep(interval)
-        elapsed += interval
-
-    print(f"Timeout after {timeout}s waiting for SSH connectivity to {address}:{ssh_port}", file=sys.stderr)
-    return False
-
-
-def run_shell_cmd(command, dry_run=False):
-    """Run a shell command and return (returncode, stdout, stderr).
-
-    Args:
-        command: list of command arguments
-        dry_run: if True, print the command instead of executing
-
-    Returns:
-        (returncode, stdout, stderr) tuple
-    """
-    if dry_run:
-        print(f"[dry-run] {' '.join(command)}")
-        return 0, "", ""
-
-    try:
-        result = subprocess.run(command, capture_output=True, text=True)
-        return result.returncode, result.stdout, result.stderr
-    except FileNotFoundError:
-        print(f"Error: '{command[0]}' not found. Is it installed and on PATH?", file=sys.stderr)
-        return 1, "", f"'{command[0]}' not found"
+from deplodock.provisioning.ssh import wait_for_ssh
+from deplodock.provisioning.shell import run_shell_cmd
 
 
 def register_vm_command(subparsers):
